moved section creation into the section model
